refactoring core (switched datasource and core, added model)
