+++ dependencynet/core/datasource/encoders.py
