+++ dependencynet/core/datasource/resourcesloader.py
