+++ dependencynet/core/model/__init__.py
@@ -1,0 +1,3 @@
+# __init__.py
+import logging
+logger = logging.getLogger(__name__)
