+++ tests/core/model/__init__.py
@@ -1,0 +1,1 @@
+# __init__.py
